2023/golang/7/solution/solution_test.go
package solution_test

import (
	_ "embed"
	"strings"
	"testing"

	soln "github.com/vdinovi/advent_of_code/challenges/2023/golang/7/solution"
)

//go:embed solution_test/input.txt
var input_txt string

func TestSolutionP1(t *testing.T) {
	tests := []struct {
		id       string
		input    soln.Input
		expected soln.Answer
	}{
		{
			id: "Sample 1",
			input: strings.NewReader(`32T3K 765
			T55J5 684
			KK677 28
			KTJJT 220
			QQQJA 483`),
			expected: 6440,
		},
		{
			id:       "Puzzle 1",
			input:    strings.NewReader(input_txt),
			expected: 250347426,
		},
	}

	for _, test := range tests {
		answer, err := soln.SolveP1(test.input)
		if err != nil {
			t.Errorf("unexpected error: %s", err)
			continue
		}
		if answer != test.expected {
			t.Errorf("expected test %q to yield answer %v but got %v ", test.id, test.expected, answer)
		}
	}
}

func TestSolutionP2(t *testing.T) {
	tests := []struct {
		id       string
		input    soln.Input
		expected soln.Answer
	}{
		{
			id: "Sample 1",
			input: strings.NewReader(`32T3K 765
			T55J5 684
			KK677 28
			KTJJT 220
			QQQJA 483`),
			expected: 5905,
		},
		{
			id:       "Puzzle 1",
			input:    strings.NewReader(input_txt),
			expected: 251224870,
		},
	}

	for _, test := range tests {
		answer, err := soln.SolveP2(test.input)
		if err != nil {
			t.Errorf("unexpected error: %s", err)
			continue
		}
		if answer != test.expected {
			t.Errorf("expected test %q to yield answer %v but got %v ", test.id, test.expected, answer)
		}
	}
}
